+++ src/types/onyx/Fund.ts
@@ -1,17 +1,32 @@
-import {ValueOf} from 'type-fest';
-import CONST from '../../CONST';
+type AdditionalData = {
+    isBillingCard?: boolean;
+    isP2PDebitCard?: boolean;
+};
+
+type AccountData = {
+    additionalData?: AdditionalData;
+    addressName?: string;
+    addressState?: string;
+    addressStreet?: string;
+    addressZip?: number;
+    cardMonth?: number;
+
+    /** The masked credit card number */
+    cardNumber?: string;
+
+    cardYear?: number;
+    created?: string;
+    currency?: string;
+    fundID?: number;
+};
 
 type Fund = {
-    cardID?: number;
-    state?: number;
-    bank?: string;
-    availableSpend?: number;
-    domainName?: string;
-    maskedPan?: string;
-    isVirtual?: boolean;
-    fraud?: ValueOf<typeof CONST.EXPENSIFY_CARD.FRAUD_TYPES>;
-    cardholderFirstName?: string;
-    cardholderLastName?: string;
+    accountData?: AccountData;
+    accountType?: string;
+    description?: string;
+    key?: string;
+    methodID?: number;
+    title?: string;
 };
 
 export default Fund;
